backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
requests==2.31.0
twilio==8.10.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
faker==20.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
drf-spectacular==0.26.5
django-ratelimit==4.1.0
python-magic==0.4.27
moviepy==1.0.3 